+++ core/src/com/coronadefense/receiver/messages/FightRoundMessage.kt
@@ -1,5 +1,5 @@
 package com.coronadefense.receiver.messages
 
 data class FightRoundMessage(
-    val roundNumber: UShort,
+    val roundNumber: Int,
 ): IMessage
