+++ core/src/com/coronadefense/receiver/messages/HealthUpdateMessage.kt
@@ -1,5 +1,5 @@
 package com.coronadefense.receiver.messages
 
 data class HealthUpdateMessage(
-    val newValue: UShort,
+    val newValue: Int,
 ): IMessage
