+++ core/src/com/coronadefense/receiver/messages/InputRoundMessage.kt
@@ -1,5 +1,5 @@
 package com.coronadefense.receiver.messages
 
 data class InputRoundMessage(
-  val roundNumber: UShort,
+  val roundNumber: Int,
 ): IMessage
