+++ core/src/com/coronadefense/receiver/messages/PlayerCountUpdateMessage.kt
@@ -1,5 +1,5 @@
 package com.coronadefense.receiver.messages
 
 data class PlayerCountUpdateMessage(
-    val playerCount: UByte, 
+    val playerCount: Int,
 ): IMessage
