+++ core/src/com/coronadefense/receiver/messages/TowerPositionMessage.kt
@@ -1,8 +1,8 @@
 package com.coronadefense.receiver.messages
 
 data class TowerPositionMessage(
-    val towerId: UShort,
-    val typeNumber: UByte,
-    val xPosition: UByte,
-    val yPosition: UByte,
+    val towerId: Int,
+    val typeNumber: Int,
+    val xPosition: Int,
+    val yPosition: Int,
 ): IMessage
